fix: fix the bug that step_preprocess report
`step_preprocess()` inserts logging messages into the report in commit c261c02b. This leaded to two problems: 1. The messages are noisy and messy. 2. (serious) it leads to `rmarkdown::render()` blockage due to unknown reasons.

To solve these problems, we first surround the logging message with "<AI></AI>". Then in `polish_report()`, if AI polish is on, the content is feed to LLM to process. Otherwise, the content in "<AI></AI>" is removed to ensure smooth report generation.